fly deployment/requirements.txt
asyncio
nest_asyncio
pandas
python-telegram-bot
pytz
apscheduler
subprocess-run
telethon
openai
argparse
pymorphy2
scikit-learn